Modularization by layers and project resctructuration
