+++ app/src/main/java/com/example/interrapidismotest/data/database/Table.kt
@@ -1,4 +1,4 @@
-package com.example.interrapidismotest.dataapp.database
+package com.example.interrapidismotest.data.database
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
