+++ app/src/main/java/com/example/interrapidismotest/ui/main/MainActivity.kt
@@ -2,12 +2,7 @@ package com.example.interrapidismotest.ui.main
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import com.example.interrapidismotest.R
-import com.example.interrapidismotest.data.datasource.TableRemoteDataSource
-import com.example.interrapidismotest.data.repository.TableRepository
-import com.example.interrapidismotest.dataapp.server.TableRemoteDataSourceImpl
 import com.example.interrapidismotest.databinding.ActivityMainBinding
-import com.example.interrapidismotest.usecases.tables.RequestTablesUseCase
 
 class MainActivity : AppCompatActivity() {
