+++ app/src/main/java/com/example/interrapidismotest/ui/tablesmenu/MenuFragment.kt
@@ -8,8 +8,8 @@ import androidx.fragment.app.viewModels
 import com.example.interrapidismotest.App
 import com.example.interrapidismotest.R
 import com.example.interrapidismotest.data.repository.TableRepository
-import com.example.interrapidismotest.dataapp.database.TableLocalDataSourceImpl
-import com.example.interrapidismotest.dataapp.server.TableRemoteDataSourceImpl
+import com.example.interrapidismotest.data.database.TableLocalDataSourceImpl
+import com.example.interrapidismotest.data.server.TableRemoteDataSourceImpl
 import com.example.interrapidismotest.databinding.FragmentMenuBinding
 import com.example.interrapidismotest.ui.common.TablesNamesAdapter
 import com.example.interrapidismotest.ui.common.diff
@@ -18,8 +18,8 @@ import com.example.interrapidismotest.ui.common.setVisibleOrGone
 import com.example.interrapidismotest.ui.common.showErrorSnackBar
 import com.example.interrapidismotest.ui.common.toggleText
 import com.example.interrapidismotest.ui.common.toggleVisibility
-import com.example.interrapidismotest.usecases.tables.GetTablesUseCase
-import com.example.interrapidismotest.usecases.tables.RequestTablesUseCase
+import com.example.interrapidismotest.usecases.GetTablesUseCase
+import com.example.interrapidismotest.usecases.RequestTablesUseCase
 
 
 class MenuFragment : Fragment(R.layout.fragment_menu) {
