+++ build.gradle
@@ -2,5 +2,7 @@
 plugins {
     id 'com.android.application' version '8.0.2' apply false
     id 'com.android.library' version '8.0.2' apply false
-    id 'org.jetbrains.kotlin.android' version '1.8.20' apply false
+    id 'org.jetbrains.kotlin.android' version '1.8.0' apply false
+    id 'com.google.devtools.ksp' version '1.8.10-1.0.9' apply false
+    id 'org.jetbrains.kotlin.jvm' version '1.8.10' apply false
 }
