+++ data/src/main/java/com/example/interrapidismotest/data/repository/TableRepository.kt
@@ -3,8 +3,6 @@ package com.example.interrapidismotest.data.repository
 import com.example.interrapidismotest.data.datasource.TableLocalDataSource
 import com.example.interrapidismotest.data.datasource.TableRemoteDataSource
 import com.example.interrapidismotest.domain.Error
-import com.example.interrapidismotest.domain.Table
-import kotlinx.coroutines.flow.Flow
 
 class TableRepository(
     private val localDataSource: TableLocalDataSource,
@@ -12,8 +10,6 @@ class TableRepository(
 ) {
 
     val tables get() = localDataSource.tables
-
-    fun findTableByName(tableName: String): Flow<Table> = localDataSource.findTableByName(tableName)
 
     suspend fun requestTablesData(): Error? {
         if (localDataSource.isEmpty()) {
