+++ domain/src/main/java/com/example/interrapidismotest/domain/Error.kt
