+++ domain/src/main/java/com/example/interrapidismotest/domain/Table.kt
