+++ settings.gradle
@@ -14,3 +14,6 @@ dependencyResolutionManagement {
 }
 rootProject.name = "InterRapidismoTest"
 include ':app'
+include ':domain'
+include ':usecases'
+include ':data'
